Credit Card Financial Dashboard Dataset
This dataset, used for building the Credit Card Financial Dashboard, contains data on credit card transactions, customer demographics, balances, and delinquency rates. It served as the foundation for analyzing and visualizing key metrics, such as total transactions, spending trends, and customer behavior, in the Power BI dashboard. The dataset highlights practical data management and analysis techniques, with SQL used for querying and transforming the data.